% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{discover}
\alias{discover}
\title{Search engines discovery schema}
\usage{
discover(ids)
}
\arguments{
\item{ids}{a character vertor of event identifier(s)}
}
\value{
a list of the event schema
}
\description{
Search engines discovery schema
}
\examples{
discover(c("R-HSA-446203", "R-HSA-73893"))
}
