% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.id.R
\name{event2Ids}
\alias{event2Ids}
\title{Event to non-Reactome ids}
\usage{
event2Ids(event.id)
}
\arguments{
\item{event.id}{a stable or db id of an Event (Pathways and Reactions)}
}
\value{
a list containing all non-Reactome identifiers associated with the given Event
}
\description{
Event to non-Reactome ids
}
\examples{
event2Ids("69541")
}
\seealso{
\code{\link{getParticipants}} for all entities of an Event
}
