% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{exportEvent}
\alias{exportEvent}
\title{Format exporter for events}
\usage{
exportEvent(event.id, format = c("sbgn", "sbml"))
}
\arguments{
\item{event.id}{a stable or db id of a pathway or reaction}

\item{format}{either in sbgn (Systems Biology Graphical Notation) or sbml (Systems Biology Markup Language)}
}
\value{
content of sbgn or sbml
}
\description{
Format exporter for events
}
\examples{
exportEvent("R-HSA-68616", "sbml")
}
