% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{exportEventFile}
\alias{exportEventFile}
\title{File exporter for events}
\usage{
exportEventFile(event.id, format = c("sbgn", "sbml"), file = NULL)
}
\arguments{
\item{event.id}{a stable or db id of a pathway or reaction}

\item{format}{either in "sbgn" (SBGN, Systems Biology Graphical Notation) or "sbml" (SBML, Systems Biology Markup Language)}

\item{file}{full path of the output file}
}
\value{
content of SBGN or SBML or a file saved into specified path
}
\description{
File exporter for events
}
\examples{
exportEventFile("R-HSA-432047", "sbgn")
# exportEventFile("R-HSA-68616", "sbml", "R-HSA-68616.sbml.xml")
}
