% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{exportImage}
\alias{exportImage}
\title{Image exporter}
\usage{
exportImage(
  id = NULL,
  output = c("diagram", "fireworks", "reaction"),
  species = NULL,
  format = c("png", "jpg", "jpeg", "svg", "gif"),
  quality = 5,
  flg = NULL,
  flgInteractors = TRUE,
  sel = NULL,
  title = TRUE,
  margin = 15,
  ehld = TRUE,
  diagramProfile = "Modern",
  token = NULL,
  resource = "TOTAL",
  analysisProfile = "Standard",
  expColumn = NULL,
  fireworksCoverage = FALSE,
  file = NULL,
  ...
)
}
\arguments{
\item{id}{stable or db id of a ReactionLikeEvent for "reaction" output, or id of an Event for "diagram"}

\item{output}{type of exported image including "diagram", "fireworks", "reaction"}

\item{species}{name or db id or taxon id of a species. Used in "fireworks" output}

\item{format}{output format including "png", "jpg", "jpeg", "svg", "gif"}

\item{quality}{result image quality between 1-10, default is 5}

\item{flg}{gene name, protein or chemical identifier or Reactome identifier used to flag elements in the diagram}

\item{flgInteractors}{defines whether to take into account interactors for the flagging}

\item{sel}{highlight element(s) selection in the diagram}

\item{title}{whether the name of the pathway is shown below}

\item{margin}{defines the image margin between 0-20, default is 15}

\item{ehld}{whether textbook-like illustration are taken into account}

\item{diagramProfile}{color profile, "modern" or "standard"}

\item{token}{token from Reactome \href{https://reactome.org/dev/analysis}{Analysis Service}}

\item{resource}{the analysis resource for which the results will be overlaid on top of the given pathways overview}

\item{analysisProfile}{analysis color profile including "Standard", "Strosobar", "Copper\%20plus"}

\item{expColumn}{the specific expression analysis results column to be
overlaid. If it is not specified (null), the first one is selected. If it is
not specified (null) and format is gif, then an animated gif is generated
with all the columns.}

\item{fireworksCoverage}{set \code{TRUE} to overlay analysis coverage values in a fireworks image}

\item{file}{full path of the output file}

\item{...}{additional parameters passed to \code{\link[magick:editing]{magick::image_write()}}}
}
\value{
an image saved into the specified path or a magick image object.
More magick processing see the \href{https://docs.ropensci.org/magick}{package}.
}
\description{
The diagram exporter allows researchers to include images of their favorite
pathway diagrams into their publications, posters or presentations. For details
see Reactome \href{https://reactome.org/dev/content-service/diagram-exporter}{diagram exporter} guide.
}
\examples{
# fireworks
exportImage(species="9606", output="fireworks", 
            format="jpg", quality=7)

\dontrun{
# animated gifs of EHLDs
# can use your own token from your Reactome Analysis
gif <- exportImage(id="R-HSA-69278", output="diagram", format="gif",
                   sel="R-HSA-69242", token="MjAyMDA2MTcyMDM5NDBfMzU2")
print(gif)

# reaction
exportImage(id="R-HSA-6787403", output="reaction", format="svg",
       flg="MTO1", analysisProfile="Copper\%20plus", file="R-HSA-6787403.svg")

}
}
\seealso{
\link{magick} to further process the image object

Other exporter: 
\code{\link{exportEventFile}()}
}
\concept{exporter}
