% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getEventsHierarchy}
\alias{getEventsHierarchy}
\title{EventsHierarchy queries}
\usage{
getEventsHierarchy(species)
}
\arguments{
\item{species}{species name or taxon id}
}
\value{
a nested dataframe containing full event hierarchy for any given species
}
\description{
EventsHierarchy queries
}
\examples{
getEventsHierarchy("Caenorhabditis elegans")
}
