% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getMapping}
\alias{getMapping}
\title{Mapping related queries}
\usage{
getMapping(id, resource, species, mapTo = c("pathways", "reactions"))
}
\arguments{
\item{id}{a stable or db id of an event or entity}

\item{resource}{database name other than Reactome (e.g. UniProt, GeneCards)}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{mapTo}{retrieve pathways or reactions where an identifier can be mapped to}
}
\value{
a dataframe containing requested pathways or reactions
}
\description{
Mapping related queries
}
\examples{
getMapping("Q7Z569", "GeneCards", "human", "reactions")
}
