% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getOrthology}
\alias{getOrthology}
\title{Orthology related queries}
\usage{
getOrthology(id, species.id)
}
\arguments{
\item{id}{a stable or db id of an event or entity}

\item{species.id}{the taxon id for a species}
}
\value{
a dataframe containing requested participants
}
\description{
Orthology related queries
}
\examples{
getOrthology("R-HSA-5674003", "49633")
}
