% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getParticipants}
\alias{getParticipants}
\title{Participants queries}
\usage{
getParticipants(
  event.id,
  retrieval = c("AllInstances", "PhysicalEntities", "ReferenceEntities",
    "EventsInPathways")
)
}
\arguments{
\item{event.id}{a stable or db id of an Event (pathways and reactions)}

\item{retrieval}{to retrieve all participants or PhysicalEntities or ReferenceEntities in an Event, or Events in a Pathway}
}
\value{
a dataframe containing requested participants
}
\description{
Participants queries
}
\examples{
getParticipants("R-HSA-6804741", "AllInstances")
# getParticipants("R-HSA-69306", "EventsInPathways")
# getParticipants("R-HSA-5205685", "PhysicalEntities")
}
