% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getPathways}
\alias{getPathways}
\title{Pathway related queries}
\usage{
getPathways(id, species = NULL, allForms = FALSE, top.level = FALSE)
}
\arguments{
\item{id}{a stable or db id of a PhysicalEntity or Event present in the pathways}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{allForms}{if set to \code{TRUE}, all low level pathways that contain the given PhysicalEntity (not Event) in all forms returned}

\item{top.level}{if set to \code{TRUE}, only top-level pathways returned}
}
\value{
a dataframe containing requested pathways
}
\description{
To get the Events that contain the given PhysicalEntity or Event (i.e. subpathway).
}
\examples{
getPathways("R-HSA-199420", "Homo sapiens")
}
\seealso{
Other getContent: 
\code{\link{getEntities}()},
\code{\link{getEventsHierarchy}()},
\code{\link{getOrthology}()},
\code{\link{getParticipants}()},
\code{\link{getPerson}()},
\code{\link{getSchemaClass}()},
\code{\link{getSpecies}()},
\code{\link{query}()}
}
\concept{getContent}
