% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getPerson}
\alias{getPerson}
\title{Person queries}
\usage{
getPerson(name = NULL, id = NULL, attributes = NULL)
}
\arguments{
\item{name}{Person’s first or/and last name}

\item{id}{Person's Orcid Id or DbId}

\item{attributes}{Property for a person. Return all available attributes if it is not specified.}
}
\value{
a list of requested information
}
\description{
Person queries
}
\examples{
getPerson(name="Robin Haw", attributes=c("displayName", "affiliation"))
}
