% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getReferences}
\alias{getReferences}
\title{ReferenceEntity queries}
\usage{
getReferences(external.id)
}
\arguments{
\item{external.id}{an id from external dabatases, e.g. ChEBI, UniProt}
}
\value{
a dataframe containing all reference entities for a given id
}
\description{
ReferenceEntity queries
}
\examples{
getReferences("15377")
}
