% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getSpecies}
\alias{getSpecies}
\title{Species queries}
\usage{
getSpecies(main = FALSE)
}
\arguments{
\item{main}{determine whether return main species, which are those have either manually curated or computationally inferred pathways}
}
\value{
a dataframe of species information
}
\description{
This method retrieves the list of all or main species in Reactome knowledgebase.
}
\examples{
# get a list of main species
getSpecies(main=TRUE)
}
\seealso{
Other getContent: 
\code{\link{getEntities}()},
\code{\link{getEventsHierarchy}()},
\code{\link{getOrthology}()},
\code{\link{getParticipants}()},
\code{\link{getPathways}()},
\code{\link{getPerson}()},
\code{\link{getSchemaClass}()},
\code{\link{query}()}
}
\concept{getContent}
