% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.id.R
\name{map2Events}
\alias{map2Events}
\title{Non-Reactome id to Events}
\usage{
map2Events(id, resource, species, mapTo = c("pathways", "reactions"))
}
\arguments{
\item{id}{non-Reactome identifier}

\item{resource}{database name other than Reactome (e.g. UniProt, GeneCards)}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{mapTo}{retrieve pathways or reactions where an identifier can be mapped to}
}
\value{
a dataframe containing requested pathways or reactions
}
\description{
Non-Reactome id to Events
}
\examples{
map2Events("Q7Z569", "GeneCards", "human", "reactions")
}
