% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.id.R
\name{map2RefEntities}
\alias{map2RefEntities}
\title{Non-Reactome id to ReferenceEntity}
\usage{
map2RefEntities(external.id)
}
\arguments{
\item{external.id}{an id from external dabatases, e.g. ChEBI, UniProt}
}
\value{
a list containing all ReferenceEntities for a given id
}
\description{
Non-Reactome id to ReferenceEntity
}
\examples{
map2RefEntities("15377") #ChEBI id
}
