% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.id.R
\name{nonReactomeId}
\alias{nonReactomeId}
\alias{map2RefEntities}
\alias{map2Events}
\alias{event2Ids}
\title{Non-Reactome id mappings}
\usage{
map2RefEntities(external.id)

map2Events(external.id, resource, species, mapTo = c("pathways", "reactions"))

event2Ids(event.id)
}
\arguments{
\item{external.id}{an id from external dabatases, e.g. ChEBI, UniProt}

\item{resource}{database name other than Reactome (e.g. UniProt, GeneCards)}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{mapTo}{retrieve Pathways or Reactions where an identifier can be mapped to}

\item{event.id}{a stable or db id of an Event (Pathways and Reactions)}
}
\value{
a list of ReferenceEntity information

a dataframe of Events associated with the given id

non-Reactome ids associated with a given Event
}
\description{
Mappings between non-Reactome identifiers and Reactome objects.
}
\details{
Descriptions of functions:
\itemize{
\item \link{map2RefEntities} maps a non-Reactome id to ReferenceEntities that store the given external id
\item \link{map2Events} fetches Events related to a given non-Reactome id
\item \link{event2Ids} gets all non-Reactome ids associated with a given Event
}
}
\examples{
\dontrun{
map2RefEntities("15377") #ChEBI id
}
map2Events("Q7Z569", resource="GeneCards", species="human", mapTo="reactions")
event2Ids("R-HSA-69541")
}
\concept{nonReactomeId}
