% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome4r-package.R
\docType{package}
\name{reactome4r-package}
\alias{reactome4r}
\alias{reactome4r-package}
\title{reactome4r: R Interface for Reactome data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chi Lam Poon \email{clpoon807@gmail.com}

}
\keyword{internal}
