% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome4r-package.R
\docType{package}
\name{ReactomeContentService4R-package}
\alias{ReactomeContentService4R}
\alias{ReactomeContentService4R-package}
\title{ReactomeContentService4R: R Interface for Reactome Content Service}
\description{
Reactome is a free, open-source, open access, curated and peer-reviewed knowledgebase of 
    biomolecular pathways. ReactomeContentService4R provides an interface to query data from the Reactome 
    ContentService. Pre-built functions in the package would allow users to retrieve data such as proteins, 
    pathways, along with other instances that are related to a specific gene or entity in Reactome.
}
\author{
\strong{Maintainer}: Reactome \email{help@reactome.org}

Authors:
\itemize{
  \item Chi-Lam Poon \email{clpoon807@gmail.com} (\href{https://orcid.org/0000-0001-6298-7099}{ORCID})
}

}
\keyword{internal}
