% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spellCheck}
\alias{spellCheck}
\title{Spell-check suggestions for a given query}
\usage{
spellCheck(query)
}
\arguments{
\item{query}{a search term}
}
\value{
spell-check suggestions for a given search term
}
\description{
Spell-check suggestions for a given query
}
\examples{
spellCheck("intelukin")
}
