% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{discover}
\alias{discover}
\title{Search engines discovery schema}
\usage{
discover(event.id)
}
\arguments{
\item{event.id}{an event identifier}
}
\value{
a list of the event schema
}
\description{
Search engines discovery schema
}
\examples{
discover("R-HSA-73893")
}
