% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{exportEventFile}
\alias{exportEventFile}
\title{File exporter}
\usage{
exportEventFile(
  id,
  format = c("sbgn", "sbml"),
  writeToFile = TRUE,
  file = NULL
)
}
\arguments{
\item{id}{a stable or db id of an Event (Pathway or Reaction)}

\item{format}{either in "sbgn" (SBGN, Systems Biology Graphical Notation) or
"sbml" (SBML, Systems Biology Markup Language)}

\item{writeToFile}{If set to \code{TRUE}, the returned data would be written into
a file. If \code{file=NULL}, the output file will be automatically saved into the
working directory and named based on the \code{id} and \code{format}}

\item{file}{full path of the output file}
}
\value{
a character object with the content of SBGN/SBML for a given id, or
a SBGN/SMBL file saved into the specified path. If the output is empty character
or list, please check on \href{https://reactome.org/ContentService/}{ContentService}
or contact HelpDesk \email{help@reactome.org}.
}
\description{
Export Reactome pathway diagrams in SBGN or SBML format.
}
\examples{
exportEventFile("R-HSA-432047", "sbgn", writeToFile=FALSE)
\dontrun{exportEventFile("R-HSA-68616", "sbml", file="orc.assembly.sbml")}
}
\seealso{
Other exporter: 
\code{\link{exportImage}()}
}
\concept{exporter}
