% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getEntities}
\alias{getEntities}
\title{PhysicalEntity queries}
\usage{
getEntities(
  id,
  retrieval = c("subunits", "complexes", "componentOf", "otherForms"),
  resource = "Reactome",
  subunitsExcludeStructures = FALSE
)
}
\arguments{
\item{id}{stable or db id of a Reactome physical entity, or id from resources other than Reactome}

\item{retrieval}{entities to be retrieved}

\item{resource}{resource other than Reactome, e.g. UniProt, Ensembl}

\item{subunitsExcludeStructures}{whether contained complexes and entity sets are excluded when retrieving subunits}
}
\value{
a dataframe containing requested information
}
\description{
PhysicalEntity queries
}
\examples{
getEntities("R-HSA-5674003", "subunits")
getEntities("P00533", "complexes", "UniProt")
}
