% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getEntities}
\alias{getEntities}
\title{PhysicalEntity queries}
\usage{
getEntities(
  id,
  retrieval = c("subunits", "complexes", "componentOf", "otherForms"),
  resource = "Reactome",
  subunitsExcludeStructures = TRUE
)
}
\arguments{
\item{id}{physical entity id or id from resources other than Reactome}

\item{retrieval}{entities to be retrieved}
}
\value{
a dataframe containing requested information
}
\description{
PhysicalEntity queries
}
\examples{
getEntities("R-HSA-5674003", "subunits")
getEntities("P00533", "complexes", "UniProt")
}
