% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getEventsHierarchy}
\alias{getEventsHierarchy}
\title{EventsHierarchy queries}
\usage{
getEventsHierarchy(main.species)
}
\arguments{
\item{main.species}{name or taxon/db id or abbreviation of main species, which could be searched using \code{getSpecies(main=T)}}
}
\value{
a nested dataframe containing full event hierarchy for any given species
}
\description{
EventsHierarchy queries
}
\examples{
getEventsHierarchy("chicken")
}
