% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getOrthology}
\alias{getOrthology}
\title{Orthology related queries}
\usage{
getOrthology(id, species)
}
\arguments{
\item{id}{a stable or db id of an event or entity}

\item{species}{name or taxon id or dbId or abbreviation of species}
}
\value{
a dataframe containing requested participants
}
\description{
Orthology related queries
}
\examples{
getOrthology("R-HSA-5674003", "Sus scrofa")
}
