% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getParticipants}
\alias{getParticipants}
\title{Participants queries}
\usage{
getParticipants(
  event.id,
  class = c("all", "physicalEntities", "referenceEntities")
)
}
\arguments{
\item{event.id}{a stable or db id of an Event}

\item{class}{retrieve all participants or PhysicalEntities or referenceEntities}
}
\value{
a dataframe containing requested participants
}
\description{
Participants queries
}
\examples{
getParticipants("R-HSA-5205685", "physicalEntities")
}
