% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getParticipants}
\alias{getParticipants}
\title{Participants queries}
\usage{
getParticipants(
  event.id,
  retrieval = c("ReactionLikeEventsInPathways", "AllInstances", "PhysicalEntities",
    "ReferenceEntities")
)
}
\arguments{
\item{event.id}{a stable or db id of an Event (pathways and reactions)}

\item{retrieval}{to retrieve ReactionLikeEvents or all participants or PhysicalEntities or ReferenceEntities}
}
\value{
a dataframe containing requested participants
}
\description{
Participants queries
}
\examples{
getParticipants("R-HSA-73916", "AllInstances")
# getParticipants("R-HSA-69306", "ReactionLikeEventsInPathways")
# getParticipants("R-HSA-5205685", "PhysicalEntities")
}
