% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getParticipants}
\alias{getParticipants}
\title{Participants related queries}
\usage{
getParticipants(
  event.id,
  retrieval = c("AllInstances", "PhysicalEntities", "ReferenceEntities",
    "EventsInPathways")
)
}
\arguments{
\item{event.id}{a stable or db id of an Event (pathways and reactions)}

\item{retrieval}{participants to be retrieved, including "AllInstances", "PhysicalEntities", "ReferenceEntities", "EventsInPathways"}
}
\value{
a dataframe containing requested participants
}
\description{
Data in Reactome are organized in a hierarchical manner - Pathways contain Reactions, Reactions contain PhysicalEntities.
This function is to get the participants of a given Event.
}
\details{
Details on options of \code{retrieval} arg:
\itemize{
\item \strong{AllInstances}: retrieves all participants (PhysicalEntities) from a given Event and their ReferenceEntities
\item \strong{PhysicalEntities}: retrieves all the PhysicalEntities that take part in a given Event
\item \strong{ReferenceEntities}: retrieves the ReferenceEntities for all PhysicalEntities in every constituent Pathway/Reaction
\item \strong{EventsInPathways}: recursively retrieves all the Events contained in any given Event
}
}
\examples{
getParticipants("R-HSA-6804741", "AllInstances")
getParticipants("R-HSA-69306", "EventsInPathways")
}
\seealso{
Other getContent: 
\code{\link{getEntities}()},
\code{\link{getEventsHierarchy}()},
\code{\link{getOrthology}()},
\code{\link{getPathways}()},
\code{\link{getPerson}()},
\code{\link{getSchemaClass}()},
\code{\link{getSpecies}()},
\code{\link{query}()}
}
\concept{getContent}
