% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getPathways}
\alias{getPathways}
\title{Pathway related queries}
\usage{
getPathways(id, species = NULL, allForms = FALSE)
}
\arguments{
\item{id}{a stable or db id of a PhysicalEntity or Event present in the pathways}

\item{species}{species name or taxon id}

\item{allForms}{whether to return all low level pathwyas that contain the given PhysicalEntity (not Event) in all forms}
}
\value{
a dataframe containing requested low level pathways
}
\description{
Pathway related queries
}
\examples{
getPathways("199420", "Homo sapiens")
}
