% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getPerson}
\alias{getPerson}
\title{Person queries}
\usage{
getPerson(name = NULL, id = NULL, attributes = NULL)
}
\arguments{
\item{name}{Person’s first or/and last name}

\item{id}{Person's Orcid Id or DbId}

\item{attributes}{Property for a person. Return all available attributes if it is not specified.}
}
\value{
a list of requested information
}
\description{
Retrieves a specific person’s property by his/her name or OrcidId or dbId.
}
\examples{
getPerson(name="Robin Haw", attributes=c("displayName", "affiliation"))
}
\seealso{
Other getContent: 
\code{\link{getEntities}()},
\code{\link{getEventsHierarchy}()},
\code{\link{getOrthology}()},
\code{\link{getParticipants}()},
\code{\link{getPathways}()},
\code{\link{getSchemaClass}()},
\code{\link{getSpecies}()},
\code{\link{query}()}
}
\concept{getContent}
