% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getPerson}
\alias{getPerson}
\title{Person queries}
\usage{
getPerson(
  name = NULL,
  id = NULL,
  attributes = NULL,
  host = "https://reactome.org",
  path = "ContentService/data/person"
)
}
\arguments{
\item{name}{Person’s first or/and last name}

\item{id}{Person's OrcidId or DbId}

\item{attributes}{Property for a person}

\item{host}{reactome}

\item{path}{path to PERSON}
}
\value{
a list of requested information
}
\description{
Person queries
}
\examples{
getPerson(name="Steve Jupe", attributes=c("displayName", "publications"))
}
