% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getSchemaClass}
\alias{getSchemaClass}
\title{Schema class queries}
\usage{
getSchemaClass(
  class,
  species = NULL,
  all = FALSE,
  rows = 1000,
  minimised = FALSE,
  reference = FALSE
)
}
\arguments{
\item{class}{schema class name, details see \href{https://reactome.org/content/schema/DatabaseObject}{Reactome data schema}}

\item{species}{name or taxon id or dbId or abbreviation of species. Only Event or PhysicalEntity class can specify species}

\item{all}{to return all entries or not, default is \code{FALSE}}

\item{rows}{the number of rows of entries retrieved, default is 1000}

\item{minimised}{to retrieve simplified entries (db id, stable id, displayName, type) or not, default is \code{FALSE}}

\item{reference}{to retrieve simplified reference objects (db id, external identifier,
external database name) or not, default is \code{FALSE}. Only for ReferenceEntity or ExternalOntology class}
}
\value{
a sorted dataframe containing entries that belong to the specified schema class
}
\description{
Schema class queries
}
\examples{
getSchemaClass(class="Drug", all=TRUE)
# getSchemaClass("Regulation", rows=2000, minimised=T)
# getSchemaClass("Complex", species="pig", rows=100)
}
