% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getSpecies}
\alias{getSpecies}
\title{Species queries}
\usage{
getSpecies(main = FALSE)
}
\arguments{
\item{main}{determine whether return main species, which are those have either manually curated or computationally inferred pathways}
}
\value{
a dataframe of species information
}
\description{
Species queries
}
\examples{
getSpecies(main=TRUE)
}
