% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{listSearchItems}
\alias{listSearchItems}
\title{List the whole Reactome search items}
\usage{
listSearchItems(
  items = c("all", "species", "types", "compartments", "keywords"),
  facet = FALSE
)
}
\arguments{
\item{items}{categories of query, including "species", "types", "compartments", "keywords", or "all"}

\item{facet}{return faceting information or not}
}
\value{
available search items
}
\description{
List the whole Reactome search items
}
\examples{
listSearchItems(items=c("species", "keyword"))
}
