% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{listSearchItems}
\alias{listSearchItems}
\title{List the whole Reactome search items (species, types, compartments, keywords)}
\usage{
listSearchItems(
  items = c("all", "species", "type", "compartment", "keyword"),
  facet = FALSE
)
}
\arguments{
\item{items}{categories of query}

\item{facet}{return faceting information or not}
}
\value{
available search items
}
\description{
List the whole Reactome search items (species, types, compartments, keywords)
}
\examples{
listSearchItems(items=c("species", "keyword"))
}
