% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{query}
\alias{query}
\title{Common data retrieval}
\usage{
query(id)
}
\arguments{
\item{id}{a Reactome stable or db id}
}
\value{
a dataframe containing all reference entities for a given id
}
\description{
Common data retrieval
}
\examples{
query("R-HSA-60140")
}
