% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome4r-package.R
\docType{package}
\name{ReactomeContentService4R-package}
\alias{ReactomeContentService4R}
\alias{ReactomeContentService4R-package}
\title{ReactomeContentService4R: R Interface for the Reactome Content Service}
\description{
Reactome is a free, open-source, open access, curated and 
    peer-reviewed knowledgebase of bio-molecular pathways. This package is to 
    interact with the Reactome Content Service API. Pre-built functions would 
    allow users to retrieve data and images that consist of proteins, pathways, 
    and other molecules related to a specific gene or entity in Reactome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reactome/ReactomeContentService4R}
  \item Report bugs at \url{https://github.com/reactome/ReactomeContentService4R/issues}
}

}
\author{
\strong{Maintainer}: Chi-Lam Poon \email{clpoon807@gmail.com} (\href{https://orcid.org/0000-0001-6298-7099}{ORCID})

Other contributors:
\itemize{
  \item Reactome \email{help@reactome.org} [copyright holder]
}

}
\keyword{internal}
