% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{searchQuery}
\alias{searchQuery}
\title{Search query}
\usage{
searchQuery(
  query,
  species = NULL,
  types = NULL,
  compartments = NULL,
  keywords = NULL,
  cluster = TRUE,
  range = NULL
)
}
\arguments{
\item{query}{name or dbId or stId of a search term from any class}

\item{species}{name or taxon id or dbId or abbreviation of species}

\item{types}{type filter, such as "Protein", "Complex", "Reaction", etc}

\item{compartments}{compartment filter, such as "cytosol", "plasma membrane", "nucleoplasm", etc}

\item{keywords}{keyword filter, such as "binds", "phosphorylates", "transports", etc}

\item{cluster}{cluster returned data or not}

\item{range}{start row and the number of rows to include, e.g. \code{range = c(0, 2)}}
}
\value{
a list of information about the search term
}
\description{
Search query
}
\examples{
searchQuery(query="Biological oxidation", species="Mus musculus", types=c("Pathway", "Reaction"))
}
\seealso{
\code{\link{listSearchItems}} for available filters
}
