% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{searchQuery}
\alias{searchQuery}
\title{Search query}
\usage{
searchQuery(
  query,
  filters = c(species = "", types = "", compartments = "", keywords = ""),
  cluster = TRUE,
  range = NULL
)
}
\arguments{
\item{query}{search term}

\item{filters}{filter conditions}

\item{cluster}{cluster returned data or not}

\item{range}{start row and the number of rows to include}
}
\value{
a list of information about the search term
}
\description{
Search query
}
\examples{
searchQuery(query="Biological oxidation", 
filters=c(species="Mus musculus", types="", compartments="", keywords=""), 
range=c(0,20))
}
