% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spellCheck}
\alias{spellCheck}
\title{Spell check the given search term}
\usage{
spellCheck(query)
}
\arguments{
\item{query}{A query term}
}
\value{
a vector of spell-check suggestions
}
\description{
Spell check the given search term
}
\examples{
spellCheck("R-HSA-12355")
spellCheck("oxidition")
}
